Do not share cookies between sub-domains +
On Heroku apps live under the herokuapp.com domain, but that domain is
on the Public Suffix List; browsers check the Public Suffix List when
storing cookies, if a cookie is requested to be stored for a domain on
the list the browser will ignore the request. More info can be found in
Heroku's doc https://devcenter.heroku.com/articles/cookies-and-herokuapp-com.

By removing the domain portion of the cookie, the cookie will be stored
only for the exact domain requested.